:sparkles: Hide anonymous classes in the report
Dependencies of anonymous classes are difficult to understand since
the bytecode class name is only a dollar character and a number. There's
no simple mapping from this name to the location in the source code.
Moreover, anonymous classes are heavily used in functional languages
like Kotlin and will therefore pollute the generated HTML report.

For this reason, decycle will hide anonymous classes in the report.
Any dependency will be treated as if it belongs to the parent class.

Example:
a dependency `Foo$1 -> Bar` will now be displayed as `Foo -> Bar`